Split behavior extends diagnostics
