+++ tests/integration/frontend_diagnostics/behavior_extends/mod.rs
@@ -1,0 +1,2 @@
+mod inherited_requirements;
+mod overlaps;
